ownerclan_tool/requirements.txt
streamlit==1.41.0
requests==2.32.3
python-dotenv==1.0.1
bcrypt==4.1.2
pybase64==1.3.2
Pillow==11.0.0